Práctica de Integración sobre el ecommerce

+++ src/services/products.service.js
@@ -12,9 +12,11 @@ export const getProductById = async (productId) => {
     return await ProductModel.findById(productId)
 };
 
-export const createProduct = async (data, file) => {
+export const createProduct = async (data, file,user) => {
     const {title,description, code, price, status, stock, category} = data;
     const thumbnail = file ? `/images/${file.filename}` : null;
+
+    const owner = user.role === 'premium' ? user.email : 'adminCoder@coder.com';
 
     const newProduct = new ProductModel({
         title,
@@ -24,18 +26,24 @@ export const createProduct = async (data, file) => {
         status,
         stock,
         category,
-        thumbnail
+        thumbnail,
+        owner
     });
 
     return await newProduct.save(); 
 };
 
 
-export const updateProduct = async (productId,data,file) => {
+export const updateProduct = async (productId,data,file,user) => {
     const {title,description, code, price,status,stock,category} = data;
     const existingProduct = await ProductModel.findById(productId);
     if (!existingProduct) {
         throw new Error('Producto no encontrado');
+    }
+
+
+    if (user.role === 'premium' && existingProduct.owner !== user.email) {
+        throw new Error('No tienes permiso para actualizar este producto');
     }
 
     const thumbnail = file ? `/images/${file.filename}` : existingProduct.thumbnail;   
@@ -47,10 +55,17 @@ export const updateProduct = async (productId,data,file) => {
     );
 };
 
-export const deleteProduct = async (productId) => {
-    const deletedProduct = await ProductModel.findByIdAndDelete(productId);
-    if (!deletedProduct) {
+export const deleteProduct = async (productId, user) => {
+    const existingProduct = await ProductModel.findById(productId);
+    if (!existingProduct) {
         throw new Error('Producto no encontrado');
     }
+
+    if (user.role !== 'admin' && (user.role === 'premium' && existingProduct.owner !== user.email)) {
+        throw new Error('No tienes permiso para eliminar este producto');
+    }
+
+    const deletedProduct = await ProductModel.findByIdAndDelete(productId);
     return deletedProduct;
-}
+
+};
